chore: move summary to utils 🧹
