+++ slides/src/utils/summary.ts
